requirements.txt
# C.O.R.E. (Coordinated Oncology Readiness Engine)
# Python Dependencies

# Google AI ADK (Agent Development Kit)
google-adk>=0.1.0  # For Agent, Runner, Tool classes

# Google GenAI (underlying SDK)
google-generativeai>=0.3.0

# Environment management
python-dotenv>=1.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlite3  # Built-in with Python

# Web framework (for Streamlit UI)
streamlit>=1.28.0

# HTTP requests (for external APIs)
requests>=2.31.0

# Biopython (for PubMed API)
biopython>=1.81

# JSON handling (built-in)
# json  # Built-in with Python

# Logging and monitoring
prometheus-client>=0.18.0  # For metrics

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type hints
typing-extensions>=4.8.0

# Optional: For advanced tracing/observability
# opentelemetry-api>=1.20.0
# opentelemetry-sdk>=1.20.0

# Development tools (optional)
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0    # Type checking